feat: update models to match mysql schema
Since we're moving the database to a central mysql db, we have to adjust
the models to match the types of columns in our mysql db.

+++ src-tauri/src/utils/models.rs
@@ -14,17 +14,17 @@ pub struct ServiceData {
 
 #[derive(Serialize, Clone, FromRow, Debug)]
 pub struct FeedbackDataRow {
-    pub id: u32,
+    pub id: i32,
     pub data: String,
     pub tag: String,
-    pub created_at: u32,
+    pub created_at: i64,
 }
 
 // NOTE: we will use this to store the ID of Row
 // that we will archive
 #[derive(Serialize, Clone, FromRow, Debug)]
 pub struct ArtifactRow {
-    pub id: u32,
+    pub id: i32,
 }
 
 // NOTE: i wrote some lifetimes for the first time
